+++ src/test/java/openmap/standard/testXmlParser.java
@@ -33,20 +33,19 @@ public class testXmlParser {
      */
     @Test
     public void ShouldLoad12Nodes(){
-
-        Map<Long, Integer> nodeWaysCounter = new HashMap<>();
-        nodeWaysCounter.put(1156449100L, 1);
-        nodeWaysCounter.put(1156449059L, 2);
-        nodeWaysCounter.put(1156448987L, 1);
-        nodeWaysCounter.put(1156449064L, 1);
-        nodeWaysCounter.put(1632757147L, 1);
-        nodeWaysCounter.put(2140481886L, 1);
-        nodeWaysCounter.put(2140481884L, 1);
-        nodeWaysCounter.put(2140481883L, 1);
-        nodeWaysCounter.put(2140481887L, 1);
-        nodeWaysCounter.put(1156449038L, 1);
-        nodeWaysCounter.put(1511529774L, 1);
-        nodeWaysCounter.put(1156449155L, 1);
+        Map<Long, Byte> nodeWaysCounter = new HashMap<>();
+        nodeWaysCounter.put(1156449100L, (byte)(1));
+        nodeWaysCounter.put(1156449059L, (byte)(2));
+        nodeWaysCounter.put(1156448987L, (byte)(1));
+        nodeWaysCounter.put(1156449064L, (byte)(1));
+        nodeWaysCounter.put(1632757147L, (byte)(1));
+        nodeWaysCounter.put(2140481886L, (byte)(1));
+        nodeWaysCounter.put(2140481884L, (byte)(1));
+        nodeWaysCounter.put(2140481883L, (byte)(1));
+        nodeWaysCounter.put(2140481887L, (byte)(1));
+        nodeWaysCounter.put(1156449038L, (byte)(1));
+        nodeWaysCounter.put(1511529774L, (byte)(1));
+        nodeWaysCounter.put(1156449155L, (byte)(1));
 
         assert(parser.parseNodes(nodeWaysCounter).size() == 12);
     }
